Add login check to necessary routes

+++ server/controllers/users.js
@@ -20,7 +20,10 @@ const login = async (req, res) => {
     const match = await bcrypt.compare(password, user.password);
 
     if (user == null || !match) {
-        return res.status(400).json({ error: 'Wrong username or password' });
+        return res.status(400).json({ 
+            status: 400,
+            error: 'Wrong username or password' 
+        });
     }
     
     req.session.userId = user.user_id;
@@ -380,7 +383,12 @@ const updateUser = async (req, res) => {
     academicYear, location, funFact, bio, skills, socials } = req.body;
     
     // Checks
-    if (!firstName) {
+    if (req.session.userId !== id){
+        return res.status(401).json({
+            status: 401, 
+            error: "Unauthorized" 
+        });
+    } else if (!firstName) {
         return res.status(400).json({
             status: 400, 
             error: "Missing user's first name" 
@@ -478,6 +486,14 @@ const deleteUser = async (req, res) => {
     // Get data
     const { id } = req.params;
 
+    // Checks
+    if (req.session.userId !== id){
+        return res.status(401).json({
+            status: 401, 
+            error: "Unauthorized" 
+        });
+    }
+
     try {
         // Delete user
         await pool.query("DELETE FROM users WHERE user_id = ?", [id]);
@@ -499,7 +515,12 @@ const updateProfilePicture = async (req, res) => {
     const { id } = req.params;
 
     // Checks
-    if (!req.file) {
+    if (req.session.userId !== id){
+        return res.status(401).json({
+            status: 401, 
+            error: "Unauthorized" 
+        });
+    } else if (!req.file) {
         return res.status(400).json({
             status: 400, 
             error: "Missing image file" 
@@ -542,6 +563,14 @@ const getAccount = async (req, res) => {
     // Get data
     const { id } = req.params;
 
+    // Checks
+    if (req.session.userId !== id){
+        return res.status(401).json({
+            status: 401, 
+            error: "Unauthorized" 
+        });
+    }
+
     try {
         // Get account information
         const sql = "SELECT u.user_id, u.primary_email, u.rit_email, u.username FROM users u WHERE user_id = ?";
@@ -570,7 +599,12 @@ const updateEmail = async (req, res) => {
     const match = await bcrypt.compare(password, curPassword[0].password);
 
     // Checks
-    if (!email || !confirm || !password) {
+    if (req.session.userId !== id){
+        return res.status(401).json({
+            status: 401, 
+            error: "Unauthorized" 
+        });
+    } else if (!email || !confirm || !password) {
         return res.status(400).json({
             status: 400, 
             error: "Missing input information" 
@@ -612,7 +646,12 @@ const updateUsername = async (req, res) => {
     const match = await bcrypt.compare(password, curPassword[0].password);
 
     // Checks
-    if (!username || !confirm || !password) {
+    if (req.session.userId !== id){
+        return res.status(401).json({
+            status: 401, 
+            error: "Unauthorized" 
+        });
+    } else if (!username || !confirm || !password) {
         return res.status(400).json({
             status: 400, 
             error: "Missing input information" 
@@ -654,7 +693,12 @@ const updatePassword = async (req, res) => {
     const match = await bcrypt.compare(password, curPassword[0].password);
 
     // Checks
-    if (!newPassword || !confirm || !password) {
+    if (req.session.userId !== id){
+        return res.status(401).json({
+            status: 401, 
+            error: "Unauthorized" 
+        });
+    } else if (!newPassword || !confirm || !password) {
         return res.status(400).json({
             status: 400, 
             error: "Missing input information" 
@@ -695,6 +739,14 @@ const getMyProjects = async (req, res) => {
 
     // Get id from url 
     const { id } = req.params;
+
+    // Checks
+    if (req.session.userId !== id){
+        return res.status(401).json({
+            status: 401, 
+            error: "Unauthorized" 
+        });
+    }
 
     try {
         // Get projects' data
@@ -786,7 +838,12 @@ const updateProjectVisibility = async (req, res) => {
     const { projectId, visibility } = req.body;
 
     // Checks
-    if (!projectId || projectId < 1) {
+    if (req.session.userId !== id){
+        return res.status(401).json({
+            status: 401, 
+            error: "Unauthorized" 
+        });
+    } else if (!projectId || projectId < 1) {
         return res.status(400).json({
             status: 400, 
             error: "Missing project id" 
@@ -866,7 +923,12 @@ const addProjectFollowing = async (req, res) => {
     const { projectId } = req.body
 
     // Checks
-    if (!projectId || projectId < 1) {
+    if (req.session.userId !== id){
+        return res.status(401).json({
+            status: 401, 
+            error: "Unauthorized" 
+        });
+    } else if (!projectId || projectId < 1) {
         return res.status(400).json({
             status: 400, 
             error: "Missing project id" 
@@ -895,7 +957,12 @@ const deleteProjectFollowing = async (req, res) => {
     const { projectId } = req.body
 
     // Checks
-    if (!projectId || projectId < 1) {
+    if (req.session.userId !== id){
+        return res.status(401).json({
+            status: 401, 
+            error: "Unauthorized" 
+        });
+    } else if (!projectId || projectId < 1) {
         return res.status(400).json({
             status: 400, 
             error: "Missing project id" 
@@ -970,7 +1037,12 @@ const addUserFollowing = async (req, res) => {
     const { userId } = req.body
 
     // Checks
-    if (!userId || userId < 1) {
+    if (req.session.userId !== id){
+        return res.status(401).json({
+            status: 401, 
+            error: "Unauthorized" 
+        });
+    } else if (!userId || userId < 1) {
         return res.status(400).json({
             status: 400, 
             error: "Missing user id" 
@@ -998,7 +1070,12 @@ const deleteUserFollowing = async (req, res) => {
     const { userId } = req.body
 
     // Checks
-    if (!userId || userId < 1) {
+    if (req.session.userId !== id){
+        return res.status(401).json({
+            status: 401, 
+            error: "Unauthorized" 
+        });
+    } else if (!userId || userId < 1) {
         return res.status(400).json({
             status: 400, 
             error: "Missing user id"
